+++ Classes/StrategiaTriplica.cs
@@ -14,7 +14,7 @@ namespace Classes
                 return Giocatore.Puntata.Stai;
         }
 
-        public override int Puntata(Giocatore giocatore, int puntataMinima, int puntataBase, int Conteggio)
+        public override double Puntata(Giocatore giocatore, double puntataMinima, double puntataBase, int Conteggio)
         {
             if (giocatore.PuntataCorrente == 0)
                 giocatore.PuntataCorrente = puntataBase;
@@ -38,7 +38,7 @@ namespace Classes
                 return Giocatore.Puntata.Stai;
         }
 
-        public override int Puntata(Giocatore giocatore, int puntataMinima, int puntataBase, int Conteggio)
+        public override double Puntata(Giocatore giocatore, double puntataMinima, double puntataBase, int Conteggio)
         {
             if (giocatore.PuntataCorrente == 0)
                 giocatore.PuntataCorrente = puntataBase;
